Added basic support for client certificate authentication
Adds support for client certificate authentication for both git operations and web access. The username is set in the form of cn=username.
If a certificate is found and the common name of the certificate matches a user then that user gets set as the authenticated user. If a
certificate is not found or does not match a user the request is completed as the anonymous user. This allows for password authentication to
continue to work.